Inline hints options refactoring
